Crate wordshk_tools

A parser for words.hk (粵典)

Parses all entries marked with OK and store the results as a list of entries. This parser does not use any regular expressions, backtracking or other inefficient parsing techniques. It is powered by a library called lip that provides flexible parser combinators and supports friendly error messages.

A note on doc format: we generally put examples after a colon ‘:’.

Structs

Def

A definition of a word

Eg

An example sentence in Mandarin, Cantonese, and/or English

Entry

An entry contains some information about a word.

Variant

A variant of a [word] with [prs] (pronounciations)

Enums

AltLang

Language tags for alternative languages other than Cantonese and English

SegmentType

Two types of segments: text or link. See Segment

Functions

parse_alt_clause

Parse a clause in an alternative language

parse_clause

Parse a Clause

parse_content

Parse the content of an Entry

parse_defs

Parse a series of definitions for a word, separated by “----”

parse_dict

Parse the whole words.hk CSV database into a Dict

parse_eg

Parse an example for a word

parse_multiline_clause

Parse a multiline clause

parse_named_clause

Parse a Clause tagged in front by its name/category

parse_partial_pr_clause

Parse a partial pronunciation Clause, until the opening paren of Jyutping pronunciations

parse_partial_pr_named_clause

Parse a partial named pronunciation Clause, until the opening paren of Jyutping pronunciations

parse_pr_clause

Parse a Jyutping pronunciation clause, for Cantonese (yue) and Mandarin (zho)

parse_rich_def

Parse a rich definition

parse_simple_def

Parse a simple definition

parse_tags

Parse tags on a word like pos, label, and sim

to_apple_dict

Convert a Dict to Apple Dictionary XML format

Type Definitions

AltClause

A clause in an alternative language other than Cantonese and English

Clause

A clause consists of one or more Lines. Appears in explanations and example sentences

Dict

A dictionary is a list of entries

Line

A line consists of one or more Segments

PrClause

An example sentence with optional Jyutping pronunciation

Segment

A segment can be a text or a link